+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/TaskCreation.kt
@@ -125,6 +125,10 @@ private fun BaseKtLintCheckTask.configureBaseCheckTask(
     android.set(pluginHolder.extension.android)
     loadedReporters.set(pluginHolder.loadReportersTask.get().loadedReporters)
     enableExperimentalRules.set(pluginHolder.extension.enableExperimentalRules)
+    projectDirectory.set(pluginHolder.target.layout.projectDirectory)
+    editorConfigFiles.from(
+        getEditorConfigFiles(pluginHolder.target.layout.projectDirectory.asFile.toPath())
+    )
 
     additionalTaskConfig()
 }