+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - Kotlin 2.4.0 support [#1059](https://github.com/JLLeitschuh/ktlint-gradle/pull/1059) (No changes needed, just updated tests)
 - Clean up workaround for Gradle 6 [#1068](https://github.com/JLLeitschuh/ktlint-gradle/pull/1068)
+- Fix build cache relocatability when using Baseline by storing relative paths in `LintErrorResult` [#1074](https://github.com/JLLeitschuh/ktlint-gradle/pull/1074)
 
 ## [14.2.0] - 2026-03-12